apt-worker: nuke mydebSystem class
This was used for special locking behaviour, so we need to see how we
can replicate / replace this, but the debSystem symbol is no longer
exported, so we have to find another way.

Honestly, I don't think this is necessary anymore. Without this commit
all the locking still seems to be done alright.